requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cycler==0.11.0
daphne==4.0.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
django-crontab==0.7.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-gis==0.18
drf-yasg==1.21.5
fonttools==4.39.4
frozenlist==1.4.0
GDAL==3.7.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.3.1
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-nmap==0.7.1
pytz==2023.3
pyzbar==0.1.9
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scapy==2.5.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
xmltodict==0.13.0
yarl==1.9.2
zope.interface==6.0
